requirements.txt
fastapi
uvicorn
pillow
pygments
jinja2
python-multipart